refactor: user 관련 패키지 구조를 레이어드 아키텍처로 변경
