+++ apps/commerce-api/src/main/java/com/loopers/domain/user/UserRepository.java
@@ -1,0 +1,12 @@
+package com.loopers.domain.user;
+
+import java.util.Optional;
+
+public interface UserRepository {
+
+    Optional<User> findByUserId(String id);
+
+    User save(User user);
+
+    Optional<User> findById(Long userId);
+}
